refactor: 템플릿 메서드 패턴을 적용해서 공통 로직을 추상 클래스로 구현한다.
