+++ app/src/main/java/com/techcourse/dao/InsertJdbcTemplate.java
@@ -1,16 +1,11 @@
 package com.techcourse.dao;
 
 import com.techcourse.domain.User;
-import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 import javax.sql.DataSource;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-public class InsertJdbcTemplate {
-
-    private static final Logger log = LoggerFactory.getLogger(UserDao.class);
+public class InsertJdbcTemplate extends JdbcTemplate {
 
     private final DataSource dataSource;
 
@@ -18,29 +13,20 @@ public class InsertJdbcTemplate {
         this.dataSource = dataSource;
     }
 
-    public void insert(User user) throws SQLException {
-        final String sql = createQueryForInsert();
-
-        Connection conn = dataSource.getConnection();
-        PreparedStatement pstmt = conn.prepareStatement(sql);
-
-        try (conn; pstmt) {
-            log.debug("query : {}", sql);
-
-            setValuesForInsert(user, pstmt);
-            pstmt.executeUpdate();
-        }
+    @Override
+    protected DataSource getDatasource() {
+        return dataSource;
     }
 
-    private String createQueryForInsert() {
-        final String sql = "insert into users (account, password, email) values (?, ?, ?)";
-        return sql;
+    @Override
+    protected String createQuery() {
+        return "insert into users (account, password, email) values (?, ?, ?)";
     }
 
-    private void setValuesForInsert(User user, PreparedStatement pstmt) throws SQLException {
+    @Override
+    protected void setValues(User user, PreparedStatement pstmt) throws SQLException {
         pstmt.setString(1, user.getAccount());
         pstmt.setString(2, user.getPassword());
         pstmt.setString(3, user.getEmail());
     }
-
 }
