+++ app/src/main/java/com/techcourse/dao/UpdateJdbcTemplate.java
@@ -1,16 +1,11 @@
 package com.techcourse.dao;
 
 import com.techcourse.domain.User;
-import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 import javax.sql.DataSource;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-public class UpdateJdbcTemplate {
-
-    private static final Logger log = LoggerFactory.getLogger(UserDao.class);
+public class UpdateJdbcTemplate extends JdbcTemplate {
 
     private final DataSource dataSource;
 
@@ -18,30 +13,21 @@ public class UpdateJdbcTemplate {
         this.dataSource = dataSource;
     }
 
-    public void update(User user) throws SQLException {
-        final String sql = createQueryForUpdate();
-
-        Connection conn = dataSource.getConnection();
-        PreparedStatement pstmt = conn.prepareStatement(sql);
-
-        try (conn; pstmt) {
-            log.debug("query : {}", sql);
-
-            setValuesForUpdate(user, pstmt);
-            pstmt.executeUpdate();
-        }
+    @Override
+    protected DataSource getDatasource() {
+        return dataSource;
     }
 
-    private String createQueryForUpdate() {
-        final String sql = "update users set account = ?, password = ?, email = ?  where id = ?";
-        return sql;
+    @Override
+    protected String createQuery() {
+        return "update users set account = ?, password = ?, email = ?  where id = ?";
     }
 
-    private void setValuesForUpdate(User user, PreparedStatement pstmt) throws SQLException {
+    @Override
+    protected void setValues(User user, PreparedStatement pstmt) throws SQLException {
         pstmt.setString(1, user.getAccount());
         pstmt.setString(2, user.getPassword());
         pstmt.setString(3, user.getEmail());
         pstmt.setLong(4, user.getId());
     }
-
 }
